requirements.txt
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.2
django-allauth==0.39.1
django-rest-auth==0.9.5
djangorestframework==3.9.4
entrypoints==0.3
idna==2.8
mccabe==0.6.1
oauthlib==3.0.1
pycodestyle==2.5.0
pyflakes==2.1.1
python-decouple==3.1
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
